# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FrozenBot API
  description: |
    API for FrozenBot - Telegram bot for frozen food delivery system.

    This API serves three main clients:
    - **Telegram Bot**: For user interactions and order processing
    - **Admin Panel**: React-based admin interface for managing orders, products, and analytics
    - **WebApp**: Telegram WebApp for customer ordering interface

    ## Authentication

    The API uses multiple authentication methods:
    - **JWT Tokens**: For admin panel authentication
    - **Telegram WebApp Validation**: For WebApp users using initData
    - **Bot Token**: For internal bot operations

    ## Rate Limiting

    API endpoints are rate limited to prevent abuse:
    - Admin endpoints: 100 requests per minute
    - User endpoints: 60 requests per minute
    - Public endpoints: 30 requests per minute

  version: 1.0.0
  contact:
    name: FrozenBot API Support
    email: support@frozenbot.com
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.frozenbot.com
    description: Production server
...
paths:
...
  /api/products/:
    get:
      tags: [products]
      summary: List products
      description: Get paginated list of products with optional filtering
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Items per page
        - name: category_id
          in: query
          schema:
            type: integer
          description: Filter by category ID
        - name: search
          in: query
          schema:
            type: string
          description: Search in product name and description
        - name: in_stock
          in: query
          schema:
            type: boolean
          description: Filter by stock availability
        - name: is_active
          in: query
          schema:
            type: boolean
            default: true
          description: Filter by active status
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

    post:
      tags: [products]
      summary: Create product
      description: Create a new product (admin only)
      security:
        - JWTAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/products/{id}:
    get:
      tags: [products]
      summary: Get product by ID
      description: Get detailed product information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Product ID
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [products]
      summary: Update product
      description: Update product information (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Product ID
      security:
        - JWTAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [products]
      summary: Delete product
      description: Delete product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Product ID
      security:
        - JWTAuth: []
      responses:
        '204':
          description: Product deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Categories
...
  /api/orders/:
    get:
      tags: [orders]
      summary: List orders
      description: Get orders (filtered by user for regular users, all orders for admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Items per page
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
          description: Filter by order status
        - name: telegram_id
          in: query
          schema:
            type: integer
            format: int64
          description: Filter by Telegram user ID (admin only)
        - name: date_from
          in: query
          schema:
            type: string
            format: date
          description: Filter orders from date (admin only)
        - name: date_to
          in: query
          schema:
            type: string
            format: date
          description: Filter orders to date (admin only)
      security:
        - JWTAuth: []
        - TelegramWebApp: []
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'

    post:
      tags: [orders]
      summary: Create order
      description: Create new order from cart
      security:
        - TelegramWebApp: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/orders/{id}:
    get:
      tags: [orders]
      summary: Get order by ID
      description: Get detailed order information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Order ID
      security:
        - JWTAuth: []
        - TelegramWebApp: []
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for admin authentication

    TelegramWebApp:
      type: apiKey
      in: header
      name: X-Telegram-Init-Data
      description: Telegram WebApp initData for user authentication
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
          additionalProperties: true
    ValidationError:
      type: object
      required:
        - error
        - message
        - details
      properties:
        error:
          type: string
          example: "validation_error"
        message:
          type: string
          example: "Validation failed"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string

    # Authentication
...
    Category:
      type: object
      required:
        - id
        - name
        - is_active
        - sort_order
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Замороженные овощи"
        description:
          type: string
          nullable: true
          example: "Свежие замороженные овощи"
        image_url:
          type: string
          nullable: true
          example: "https://example.com/category.jpg"
        is_active:
          type: boolean
          example: true
        sort_order:
          type: integer
          example: 0
        products_count:
          type: integer
          readOnly: true
          example: 15
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...
    Product:
      type: object
      required:
        - id
        - name
        - price
        - is_active
        - in_stock
        - sort_order
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Брокколи замороженная"
        description:
          type: string
          nullable: true
          example: "Свежая замороженная брокколи"
        price:
          type: number
          format: float
          example: 150.0
        formatted_price:
          type: string
          readOnly: true
          example: "150₽"
        image_url:
          type: string
          nullable: true
          example: "https://example.com/broccoli.jpg"
        is_active:
          type: boolean
          example: true
        in_stock:
          type: boolean
          example: true
        weight:
          type: integer
          nullable: true
          description: "Weight in grams"
          example: 500
        formatted_weight:
          type: string
          readOnly: true
          example: "500г"
        sort_order:
          type: integer
          example: 0
        category_id:
          type: integer
          nullable: true
          example: 1
        category:
          $ref: '#/components/schemas/Category'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
    ProductCreateRequest:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
          example: "Брокколи замороженная"
        description:
          type: string
          nullable: true
          example: "Свежая замороженная брокколи"
        price:
          type: number
          format: float
          minimum: 0
          example: 150.0
        image_url:
          type: string
          nullable: true
          example: "https://example.com/broccoli.jpg"
        is_active:
          type: boolean
          default: true
        in_stock:
          type: boolean
          default: true
        weight:
          type: integer
          nullable: true
          minimum: 0
          description: "Weight in grams"
          example: 500
        sort_order:
          type: integer
          default: 0
        category_id:
          type: integer
          nullable: true
          example: 1
    ProductUpdateRequest:
      type: object
      properties:
        name:
          type: string
          example: "Брокколи замороженная"
        description:
          type: string
          nullable: true
          example: "Свежая замороженная брокколи"
        price:
          type: number
          format: float
          minimum: 0
          example: 150.0
        image_url:
          type: string
          nullable: true
          example: "https://example.com/broccoli.jpg"
        is_active:
          type: boolean
        in_stock:
          type: boolean
        weight:
          type: integer
          nullable: true
          minimum: 0
          description: "Weight in grams"
          example: 500
        sort_order:
          type: integer
        category_id:
          type: integer
          nullable: true
          example: 1
    ProductListResponse:
      type: object
      required:
        - items
        - total
        - page
        - pages
        - per_page
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        total:
          type: integer
          description: Total number of products
          example: 150
        page:
          type: integer
          description: Current page number
          example: 1
        pages:
          type: integer
          description: Total number of pages
          example: 8
        per_page:
          type: integer
          description: Items per page
          example: 20

    # Cart
...
    OrderStatus:
      type: string
      enum:
        - pending
        - confirmed
        - preparing
        - ready
        - completed
        - cancelled
      example: "pending"
    Order:
      type: object
      required:
        - id
        - user_id
        - status
        - total_amount
        - customer_name
        - customer_phone
        - payment_method
        - items
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        status:
          $ref: '#/components/schemas/OrderStatus'
        status_display:
          type: string
          readOnly: true
          example: "Ожидает подтверждения"
        total_amount:
          type: number
          format: float
          example: 450.0
        formatted_total:
          type: string
          readOnly: true
          example: "450₽"
        customer_name:
          type: string
          example: "Иван Иванов"
        customer_phone:
          type: string
          example: "+7900123456"
        delivery_address:
          type: string
          nullable: true
          example: "ул. Пушкина, д. 1, кв. 1"
        notes:
          type: string
          nullable: true
          example: "Позвонить за 10 минут"
        payment_method:
          type: string
          example: "card"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        user:
          $ref: '#/components/schemas/User'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...
    OrderCreateRequest:
      type: object
      required:
        - telegram_id
        - customer_name
        - customer_phone
      properties:
        telegram_id:
          type: integer
          format: int64
          example: 123456789
        customer_name:
          type: string
          example: "Иван Иванов"
        customer_phone:
          type: string
          example: "+7900123456"
        delivery_address:
          type: string
          nullable: true
          example: "ул. Пушкина, д. 1, кв. 1"
        notes:
          type: string
          nullable: true
          example: "Позвонить за 10 минут"
        payment_method:
          type: string
          default: "card"
          example: "card"
...
    OrderListResponse:
      type: object
      required:
        - items
        - total
        - page
        - pages
        - per_page
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        total:
          type: integer
          description: Total number of orders
          example: 50
        page:
          type: integer
          description: Current page number
          example: 1
        pages:
          type: integer
          description: Total number of pages
          example: 3
        per_page:
          type: integer
          description: Items per page
          example: 20

    # Admin Analytics
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "bad_request"
            message: "Invalid request parameters"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "unauthorized"
            message: "Authentication required"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "forbidden"
            message: "Access denied"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "not_found"
            message: "Resource not found"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            error: "validation_error"
            message: "Validation failed"
            details:
              - field: "price"
                message: "Price must be greater than 0"
                code: "value_error.number.not_gt"